Fix style.test
