+++ js/util/dispatcher.js
@@ -35,6 +35,13 @@ function Dispatcher(length, parent) {
     workerBus.target = parentBus;
 
     this.worker = new Worker(workerBus);
+
+    // TODO: this is here temporarily to allow tests to work until the geojson
+    // worker source has been hard-coded into the Worker instead of loaded
+    // asynchronously
+    var GeoJSONWorkerSource = require('../source/geojson_worker_source.js');
+    this.worker.workerSources.geojson = new GeoJSONWorkerSource();
+
     this.actor = new Actor(parentBus, parent);
 
     this.remove = function() {
