+++ test/js/style/style.test.js
@@ -5,8 +5,9 @@ var st = require('st');
 var http = require('http');
 var path = require('path');
 var sinon = require('sinon');
+var proxyquire = require('proxyquire');
 var Style = require('../../../js/style/style');
-var VectorTileSource = require('../../../js/source/vector_tile_source');
+var SourceCache = require('../../../js/source/source_cache');
 var StyleLayer = require('../../../js/style/style_layer');
 var util = require('../../../js/util/util');
 
@@ -19,13 +20,13 @@ function createStyleJSON(properties) {
 }
 
 function createSource() {
-    return new VectorTileSource({
+    return {
         type: 'vector',
         minzoom: 1,
         maxzoom: 10,
         attribution: 'Mapbox',
         tiles: ['http://example.com/{z}/{x}/{y}.png']
-    });
+    };
 }
 
 function createGeoJSONSource() {
@@ -68,7 +69,7 @@ test('Style', function(t) {
             }
         }));
         style.on('load', function() {
-            t.ok(style.getSource('mapbox') instanceof VectorTileSource);
+            t.ok(style.getSource('mapbox') instanceof SourceCache);
             t.end();
         });
     });
@@ -90,9 +91,9 @@ test('Style', function(t) {
             style.removeSource('-source-id-');
 
             var source = createSource();
+            source['vector_layers'] = [{ id: 'green' }];
             style.addSource('-source-id-', source);
             style.update();
-            source.vectorLayerIds = ['green'];
         });
 
         style.on('error', function(event) {
@@ -277,7 +278,7 @@ test('Style#addSource', function(t) {
         var style = new Style(createStyleJSON()),
             source = createSource();
         style.on('source.add', function(e) {
-            t.equal(e.source, source);
+            t.same(e.source.serialize(), source);
             t.end();
         });
         style.on('load', function () {
@@ -325,11 +326,11 @@ test('Style#addSource', function(t) {
         var source = createSource();
 
         function sourceEvent(e) {
-            t.equal(e.source, source);
+            t.same(e.source.serialize(), source);
         }
 
         function tileEvent(e) {
-            t.equal(e.source, source);
+            t.same(e.source.serialize(), source);
         }
 
         style.on('source.load',   sourceEvent);
@@ -343,12 +344,12 @@ test('Style#addSource', function(t) {
         style.on('load', function () {
             t.plan(7);
             style.addSource('source-id', source); // Fires load
-            source.fire('error');
-            source.fire('change');
-            source.fire('tile.add');
-            source.fire('tile.load');
-            source.fire('tile.error');
-            source.fire('tile.remove');
+            style.getSource('source-id').fire('error');
+            style.getSource('source-id').fire('change');
+            style.getSource('source-id').fire('tile.add');
+            style.getSource('source-id').fire('tile.load');
+            style.getSource('source-id').fire('tile.error');
+            style.getSource('source-id').fire('tile.remove');
         });
     });
 
@@ -387,7 +388,7 @@ test('Style#removeSource', function(t) {
         var style = new Style(createStyleJSON()),
             source = createSource();
         style.on('source.remove', function(e) {
-            t.equal(e.source, source);
+            t.same(e.source.serialize(), source);
             t.end();
         });
         style.on('load', function () {
@@ -431,6 +432,8 @@ test('Style#removeSource', function(t) {
 
         style.on('load', function () {
             style.addSource('source-id', source);
+            source = style.getSource('source-id');
+
             style.removeSource('source-id');
 
             source.fire('load');
@@ -497,6 +500,7 @@ test('Style#addLayer', function(t) {
 
         style.on('load', function() {
             var source = createSource();
+            source['vector_layers'] = [{id: 'green'}];
             style.addSource('-source-id-', source);
             style.addLayer({
                 'id': '-layer-id-',
@@ -504,7 +508,6 @@ test('Style#addLayer', function(t) {
                 'source': '-source-id-',
                 'source-layer': '-source-layer-'
             });
-            source.vectorLayerIds = ['green'];
         });
 
         style.on('error', function(event) {
@@ -900,7 +903,51 @@ test('Style#setLayerZoomRange', function(t) {
 });
 
 test('Style#queryRenderedFeatures', function(t) {
-    var style = new Style({
+    var style;
+    var Style = proxyquire('../../../js/style/style', {
+        '../source/query-features': {
+            rendered: function(source, layers, queryGeom, params) {
+                if (source.id !== 'mapbox') {
+                    return [];
+                }
+
+                var features = {
+                    'land': [{
+                        type: 'Feature',
+                        layer: style._layers.land,
+                        geometry: {
+                            type: 'Polygon'
+                        }
+                    }, {
+                        type: 'Feature',
+                        layer: style._layers.land,
+                        geometry: {
+                            type: 'Point'
+                        }
+                    }],
+                    'landref': [{
+                        type: 'Feature',
+                        layer: style._layers.landref,
+                        geometry: {
+                            type: 'Line'
+                        }
+                    }]
+                };
+
+                if (params.layers) {
+                    for (var l in features) {
+                        if (params.layers.indexOf(l) < 0) {
+                            delete features[l];
+                        }
+                    }
+                }
+
+                return features;
+            }
+        }
+    });
+
+    style = new Style({
         "version": 8,
         "sources": {
             "mapbox": {
@@ -952,53 +999,6 @@ test('Style#queryRenderedFeatures', function(t) {
     style.on('load', function() {
         style._applyClasses([]);
         style._recalculate(0);
-
-        style.sources.other.queryRenderedFeatures = function(position, params) {
-            var features = {'land--other': []};
-            if (params.layers) {
-                for (var l in features) {
-                    if (params.layers.indexOf(l) < 0) {
-                        delete features[l];
-                    }
-                }
-            }
-            return features;
-        };
-
-        style.sources.mapbox.queryRenderedFeatures = function(position, params) {
-            var features = {
-                'land': [{
-                    type: 'Feature',
-                    layer: style._layers.land,
-                    geometry: {
-                        type: 'Polygon'
-                    }
-                }, {
-                    type: 'Feature',
-                    layer: style._layers.land,
-                    geometry: {
-                        type: 'Point'
-                    }
-                }],
-                'landref': [{
-                    type: 'Feature',
-                    layer: style._layers.landref,
-                    geometry: {
-                        type: 'Line'
-                    }
-                }]
-            };
-
-            if (params.layers) {
-                for (var l in features) {
-                    if (params.layers.indexOf(l) < 0) {
-                        delete features[l];
-                    }
-                }
-            }
-
-            return features;
-        };
 
         t.test('returns feature type', function(t) {
             var results = style.queryRenderedFeatures([{column: 1, row: 1, zoom: 1}], {}, 0, 0);
